refactor : profile 및 post ts-error 해결
