+++ src/components/profile/Selectkeyword.tsx
@@ -2,7 +2,7 @@ import tw from 'tailwind-styled-components';
 import Modal from '@components/common/Modal';
 import Navigate from '@components/common/Navigate';
 import { useState } from 'react';
-import Button from '../../components/common/Button';
+import Button from '../common/Button';
 import { useRouter } from 'next/router';
 
 interface KeywordForm {
