+++ src/pages/profile/keyword.tsx
@@ -1,8 +1,6 @@
 import Layout from '@components/common/Layout';
-import Modal from '@components/common/Modal';
-import { useRouter } from 'next/router';
 import { useState } from 'react';
-import SelectKeyword from '@components/common/Selectkeyword';
+import SelectKeyword from '@components/profile/Selectkeyword';
 import authApi from '@apis/auth/authApi';
 
 export default function Keyword() {
@@ -10,11 +8,11 @@ export default function Keyword() {
   const handleSubmit = async (e: { target: { id: string } }) => {
     const formData = new FormData();
     if (e.target.id === 'skip') {
-      formData.append('keywords', []);
+      formData.append('keywords', JSON.stringify([]));
     } else {
-      formData.append('keywords', keywordList);
+      formData.append('keywords', JSON.stringify(keywordList));
     }
-    await authApi.patchUserInfo(formData);
+    const res = await authApi.patchUserInfo(formData);
   };
   return (
     <Layout>
